Added Quest Progression Feature
Added the progression feature to the quest system so that quests are to
be done in order. Quests are set as locked until the previous quest is
completed. Also, I added a quest completion count to the view quest
status feature, displaying the number of times the player completed each
quest.